# openapi.yaml
openapi: 3.0.3
info:
  title: Cutpoint
  version: 0.1.0
  description: >-
    Criminal-market disruption planning. Cutpoint discovers dependency edges between
    products traded in underground markets, requires analyst adjudication before any
    edge becomes a finding, scores which defender-side control change severs the most
    downstream harm, and measures the result against the customer's own loss metrics.
  x-product-slug: ai-cyber-cutpoint
  x-source: ai-in-cyber/1812.00381v2/
servers:
  - url: https://api.example.com
...
paths:
  /v1/markets:
    get:
      operationId: listMarkets
      tags: [Markets]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: language
          in: query
          schema:
            type: string
            example: ru
        - name: status
          in: query
          schema:
            type: string
            enum: [onboarding, labelling, active, held, exited]
      responses:
        "200":
          description: Monitored markets page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelopeMarket"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: registerMarket
      tags: [Markets]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketCreate"
      responses:
        "201":
          description: Market registered with an onboarding labelling budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeMarket"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /v1/markets/{marketId}:
    parameters:
      - $ref: "#/components/parameters/MarketId"
    get:
      operationId: getMarket
      tags: [Markets]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Market with onboarding budget consumption and publication state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeMarket"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/markets/{marketId}/corpora:
    parameters:
      - $ref: "#/components/parameters/MarketId"
    get:
      operationId: listCorpora
      tags: [Markets]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Corpora page with lawful basis and retention expiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelopeCorpus"
    post:
      operationId: ingestCorpus
      tags: [Markets]
      description: >-
        Ingest a licensed corpus. Rejected unless lawful basis, licence terms, and a
        retention window are supplied, and unless pseudonymisation is requested.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CorpusIngest"
      responses:
        "202":
          description: Corpus accepted for pseudonymised intake
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeCorpus"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
  /v1/markets/{marketId}/taxonomy/versions:
    parameters:
      - $ref: "#/components/parameters/MarketId"
    post:
      operationId: publishTaxonomyVersion
      tags: [Taxonomy]
      description: >-
        Publish a new taxonomy version. Existing adjudicated chains remain bound to the
        version under which they were reviewed.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaxonomyVersionCreate"
      responses:
        "201":
          description: Taxonomy version published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeTaxonomyVersion"
        "409":
          description: A labelling batch is open against the current version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /v1/annotation/batches:
    get:
      operationId: listAnnotationBatches
      tags: [Annotation]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: marketId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, complete, abandoned]
      responses:
        "200":
          description: Annotation batches page with effort against budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelopeAnnotationBatch"
    post:
      operationId: createAnnotationBatch
      tags: [Annotation]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotationBatchCreate"
      responses:
        "201":
          description: Batch created with reserved holdout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeAnnotationBatch"
...
  /v1/evidence/packages:
    post:
      operationId: sealEvidencePackage
      tags: [Evidence]
      description: >-
        Privileged action. Seals adjudicated chains and their source artefacts into a
        hash-manifested referral package and opens a custody log.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidencePackageCreate"
      responses:
        "201":
          description: Evidence package sealed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeEvidencePackage"
        "403":
          description: Caller lacks the evidence privilege
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /v1/evidence/packages/{packageId}:
    parameters:
      - $ref: "#/components/parameters/PackageId"
    get:
      operationId: getEvidencePackage
      tags: [Evidence]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Sealed package with hash manifest and custody log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeEvidencePackage"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    MarketId:
      name: marketId
      in: path
      required: true
      schema:
        type: string
...
    PackageId:
      name: packageId
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    ResponseMeta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
    Market:
      type: object
      required: [id, name, language, status]
      properties:
        id:
          type: string
        name:
          type: string
        language:
          type: string
          description: Primary forum language, used to route fluent labellers
        status:
          type: string
          enum: [onboarding, labelling, active, held, exited]
        onboardingLabelBudgetPosts:
          type: integer
          description: Committed labelling budget for this market
          example: 8500
        labelledPosts:
          type: integer
        annotationEffortSeconds:
          type: integer
        activeTaxonomyVersion:
          type: string
        publicationHeld:
          type: boolean
        firstAdjudicatedChainAt:
          type: string
          format: date-time
    MarketCreate:
      type: object
      required: [name, language, onboardingLabelBudgetPosts]
      properties:
        name:
          type: string
        language:
          type: string
        onboardingLabelBudgetPosts:
          type: integer
          minimum: 1000
        driftThresholdF1:
          type: number
          format: float
          description: Publication is held when holdout F1 falls below this value
    SourceLicence:
      type: object
      required: [lawfulBasis, licenceTerms]
      properties:
        lawfulBasis:
          type: string
          enum: [public_posting_scrape, commercial_agreement, research_agreement, law_enforcement_disclosure]
        licenceTerms:
          type: string
        redistributionPermitted:
          type: boolean
        publicIndexingPermitted:
          type: boolean
          default: false
    RetentionPolicy:
      type: object
      required: [rawRetentionDays]
      properties:
        rawRetentionDays:
          type: integer
          description: Retention for raw posts and replies
        derivedRetentionDays:
          type: integer
          description: Retention for pseudonymised labels, adjudications, and edges
        purgeOnExpiry:
          type: boolean
          default: true
    Corpus:
      type: object
      required: [id, marketId, status]
      properties:
        id:
          type: string
        marketId:
          type: string
        status:
          type: string
          enum: [accepted, pseudonymising, available, expired, purged]
        threadCount:
          type: integer
        replyCount:
          type: integer
        uniqueActorCount:
          type: integer
        coverageStart:
          type: string
          format: date
        coverageEnd:
          type: string
          format: date
        licence:
          $ref: "#/components/schemas/SourceLicence"
        retention:
          $ref: "#/components/schemas/RetentionPolicy"
        retentionExpiresAt:
          type: string
          format: date-time
    CorpusIngest:
      type: object
      required: [licence, retention, payloadReference]
      properties:
        payloadReference:
          type: string
          description: Opaque handle to the licensed archive, never inline content
        licence:
          $ref: "#/components/schemas/SourceLicence"
        retention:
          $ref: "#/components/schemas/RetentionPolicy"
        pseudonymiseActors:
          type: boolean
          default: true
        coverageStart:
          type: string
          format: date
        coverageEnd:
          type: string
          format: date
...
    ProductCategory:
      type: object
      required: [key, label]
      properties:
        key:
          type: string
          example: hacked_server
        label:
          type: string
        definition:
          type: string
        retired:
          type: boolean
          default: false
...
    TaxonomyVersion:
      type: object
      required: [id, marketId, version, publishedAt]
      properties:
        id:
          type: string
        marketId:
          type: string
        version:
          type: string
        publishedAt:
          type: string
          format: date-time
        publishedBy:
          type: string
        changeSummary:
          type: string
        categories:
          type: array
          items:
            $ref: "#/components/schemas/ProductCategory"
    TaxonomyVersionCreate:
      type: object
      required: [categories, changeSummary]
      properties:
        categories:
          type: array
          items:
            $ref: "#/components/schemas/ProductCategory"
        changeSummary:
          type: string
    AnnotationBatch:
      type: object
      required: [id, marketId, status, targetPosts]
      properties:
        id:
          type: string
        marketId:
          type: string
        taxonomyVersion:
          type: string
        status:
          type: string
          enum: [open, in_progress, complete, abandoned]
        targetPosts:
          type: integer
        labelledPosts:
          type: integer
        labelledReplies:
          type: integer
        holdoutReservedPosts:
          type: integer
        effortSeconds:
          type: integer
        budgetConsumedShare:
          type: number
          format: float
        labellerIds:
          type: array
          items:
            type: string
    AnnotationBatchCreate:
      type: object
      required: [marketId, targetPosts]
      properties:
        marketId:
          type: string
        targetPosts:
          type: integer
        holdoutReservedPosts:
          type: integer
        labellerIds:
          type: array
          items:
            type: string
...
    CustodyEvent:
      type: object
      required: [at, actor, action]
      properties:
        at:
          type: string
          format: date-time
        actor:
          type: string
        action:
          type: string
          enum: [sealed, accessed, transferred, disclosed]
        recipient:
          type: string
    EvidencePackageCreate:
      type: object
      required: [chainAdjudicationIds, purpose]
      properties:
        chainAdjudicationIds:
          type: array
          items:
            type: string
        purpose:
          type: string
          enum: [law_enforcement_referral, internal_legal, regulator_disclosure]
        recipient:
          type: string
    EvidencePackage:
      type: object
      required: [id, purpose, sealedAt, manifestHash]
      properties:
        id:
          type: string
        purpose:
          type: string
          enum: [law_enforcement_referral, internal_legal, regulator_disclosure]
        sealedAt:
          type: string
          format: date-time
        sealedBy:
          type: string
        manifestHash:
          type: string
        artefactHashes:
          type: array
          items:
            type: object
            properties:
              artefactId:
                type: string
              artefactType:
                type: string
                enum: [post, reply, adjudication, dependency_edge]
              hash:
                type: string
        custodyLog:
          type: array
          items:
            $ref: "#/components/schemas/CustodyEvent"
    DataEnvelopeMarket:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/Market"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    DataEnvelopeCorpus:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/Corpus"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    DataEnvelopeTaxonomyVersion:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/TaxonomyVersion"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    DataEnvelopeAnnotationBatch:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/AnnotationBatch"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    DataEnvelopeEvidencePackage:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/EvidencePackage"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ListEnvelopeMarket:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Market"
            nextCursor:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ListEnvelopeCorpus:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Corpus"
            nextCursor:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ListEnvelopeAnnotationBatch:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AnnotationBatch"
            nextCursor:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...