# platform/api-server/src/domains/governance/handlers/package.handlers.ts
/**
 * Handlers for Package
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: governance
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { GovernanceDomainModule } from "../dependencies/governance-ddd.dependencies.js";
import type { GetEvidencePackageInput, SealEvidencePackageInput } from "@cutpoint/services/governance";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for POST /v1/evidence/packages
 * Operation: sealEvidencePackage
 */
export async function sealEvidencePackage(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: GovernanceDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: SealEvidencePackageInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.packages.seal.execute(input);
  return reply.code(201).send(result);
}


/**
 * Handler for GET /v1/evidence/packages/{packageId}
 * Operation: getEvidencePackage
 */
export async function getEvidencePackage(
  request: FastifyRequest<{ Params: { packageId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: GovernanceDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: GetEvidencePackageInput = {
    packageId: request.params.packageId,
  };
  const result = await deps.useCases.packages.get.execute(input);
  return reply.code(200).send(result);
}