# platform/api-server/src/domains/labelling/handlers/batch.handlers.ts
/**
 * Handlers for Batch
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: labelling
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { LabellingDomainModule } from "../dependencies/labelling-ddd.dependencies.js";
import type { CreateAnnotationBatchInput, ListAnnotationBatchesInput } from "@cutpoint/services/labelling";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/annotation/batches
 * Operation: listAnnotationBatches
 */
export async function listAnnotationBatches(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], marketId?: string | string[], status?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: LabellingDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListAnnotationBatchesInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.batches.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/annotation/batches
 * Operation: createAnnotationBatch
 */
export async function createAnnotationBatch(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: LabellingDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: CreateAnnotationBatchInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.batches.create.execute(input);
  return reply.code(201).send(result);
}