# platform/api-server/src/domains/labelling/handlers/version.handlers.ts
/**
 * Handlers for Version
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: labelling
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { LabellingDomainModule } from "../dependencies/labelling-ddd.dependencies.js";
import type { PublishTaxonomyVersionInput } from "@cutpoint/services/labelling";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for POST /v1/markets/{marketId}/taxonomy/versions
 * Operation: publishTaxonomyVersion
 */
export async function publishTaxonomyVersion(
  request: FastifyRequest<{ Params: { marketId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: LabellingDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: PublishTaxonomyVersionInput = {
    marketId: request.params.marketId,
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.versions.get.execute(input);
  return reply.code(201).send(result);
}