# platform/api-server/src/domains/markets/handlers/corpora.handlers.ts
/**
 * Handlers for Corpora
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: markets
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { MarketsDomainModule } from "../dependencies/markets-ddd.dependencies.js";
import type { IngestCorpusInput, ListCorporaInput } from "@cutpoint/services/markets";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/markets/{marketId}/corpora
 * Operation: listCorpora
 */
export async function listCorpora(
  request: FastifyRequest<{ Params: { marketId: string }; Querystring: { cursor?: string | string[], limit?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: MarketsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListCorporaInput = {
    marketId: request.params.marketId,
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.corporas.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/markets/{marketId}/corpora
 * Operation: ingestCorpus
 */
export async function ingestCorpus(
  request: FastifyRequest<{ Params: { marketId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: MarketsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: IngestCorpusInput = {
    marketId: request.params.marketId,
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.corporas.get.execute(input);
  return reply.code(200).send(result);
}