# platform/api-server/src/domains/markets/handlers/market.handlers.ts
/**
 * Handlers for Market
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: markets
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { MarketsDomainModule } from "../dependencies/markets-ddd.dependencies.js";
import type { GetMarketInput, ListMarketsInput, RegisterMarketInput } from "@cutpoint/services/markets";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/markets
 * Operation: listMarkets
 */
export async function listMarkets(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], language?: string | string[], status?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: MarketsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListMarketsInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.markets.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/markets
 * Operation: registerMarket
 */
export async function registerMarket(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: MarketsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: RegisterMarketInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.markets.create.execute(input);
  return reply.code(201).send(result);
}


/**
 * Handler for GET /v1/markets/{marketId}
 * Operation: getMarket
 */
export async function getMarket(
  request: FastifyRequest<{ Params: { marketId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: MarketsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: GetMarketInput = {
    marketId: request.params.marketId,
  };
  const result = await deps.useCases.markets.get.execute(input);
  return reply.code(200).send(result);
}